+++ src/object/null.go
@@ -1,0 +1,8 @@
+package object
+
+const NULL_OBJ = "NULL"
+
+type Null struct{}
+
+func (n *Null) Type() ObjectType { return NULL_OBJ }
+func (n *Null) Inspect() string  { return "null" }
