+++ src/object/return.go
@@ -1,0 +1,10 @@
+package object
+
+const RETURN_VALUE_OBJ = "RETURN_VALUE"
+
+type ReturnValue struct {
+	Value Object
+}
+
+func (rv *ReturnValue) Type() ObjectType { return RETURN_VALUE_OBJ }
+func (rv *ReturnValue) Inspect() string  { return rv.Value.Inspect() }
