+++ src/object/string.go
@@ -1,0 +1,10 @@
+package object
+
+const STRING_OBJ = "STRING"
+
+type String struct {
+	Value string
+}
+
+func (s *String) Type() ObjectType { return STRING_OBJ }
+func (s *String) Inspect() string  { return s.Value }
